requirements.txt
python-dotenv
openai
langchain
langchain-community
langchain_anthropic
langchain-google-genai 
llama-index
openpyxl
matplotlib
pillow
datasets
scikit-learn

